requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
python-rocksdb>=0.7.0
tqdm>=4.62.0
